Final push before ISP day
One of the nixies did not work, so the seconds digits had to be
removed. cycleTubes() function more appealing and less time consuming.
Has a neat slowing down effect now. ATMega328p PWM resonated with the
caps in the MOSFET and fried the chip, as such, brightness function was
removed. The boost converter was replaced with a dedicated circuit.